sp1_primitives

Module consts

Modules

fd

Constants

BABYBEAR_PRIME

The Baby Bear prime.

BYTE_SIZE

The number of bits in a byte.

LONG_WORD_SIZE

The number of bytes necessary to represent a 64-bit integer.

MAXIMUM_MEMORY_SIZE

The maximum size of the memory in bytes.

WORD_SIZE

The size of a word in bytes.

Functions

bytes_to_words_le

Converts a byte array in little endian to a slice of words.

bytes_to_words_le_vec

Converts a byte array in little endian to a vector of words.

num_to_comma_separated

words_to_bytes_le

Converts a slice of words to a slice of bytes in little endian.

words_to_bytes_le_vec

Converts a slice of words to a byte vector in little endian.